A different approach to the reporting object.
Looking at the suggestions made my Mike Gee, I saw the possibility of a
different approach. I refactored the hash and list of words into a table
made of an array of pairs.

This cleaned up the intermediate steps, got rid of an instance variable
and focused this object on just what it needed to do: prepare a table of
popular words for printing.